ShudhiNetra Nirikshak.AI/backend/backend/post_office/serializers.py
from rest_framework import serializers
from .models import PostOffice

class PostOfficeSerializer(serializers.ModelSerializer):
    class Meta:
        model=PostOffice
        fields="__all__"
    
        