ShudhiNetra Nirikshak.AI/backend/requirements.txt
Django
django-cors-headers
djangorestframework
